pyproject.toml
[tool.poetry]
name = "hw27"
version = "0.1.0"
description = ""
authors = ["KROKS13kir <arbitr13adv@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.8"
Django = "^4.1.1"
psycopg2 = "^2.9.3"
djangorestframework = "^3.14.0"
pytest-django = "^4.5.2"
djangorestframework-simplejwt = "^5.2.2"
pytest-factoryboy = "^2.5.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
